Implemented a first attempt at a jump function. Fixed type conversion errors.

+++ game1/Vector2.cpp
@@ -6,7 +6,7 @@ Vector2::Vector2()
 	y = 0;
 }
 
-Vector2::Vector2(float X, float Y)
+Vector2::Vector2(double X, double Y)
 {
 	x = X;
 	y = Y;
@@ -17,7 +17,7 @@ Vector2 Vector2::add(Vector2 v)
 	return *new Vector2(x + v.x, y + v.y);
 }
 
-Vector2 Vector2::multiply(float n)
+Vector2 Vector2::multiply(double n)
 {
 	return *new Vector2(x * n, y * n);
 }
